requirements.txt
flask
paho-mqtt
opencv-python
